Created Flavors Table (See Notes)
Migration started for likes as foreign_key, but not migrated because it
is not clear whether or not likes is a foreign key yet since it is
just a measurement of the number of times a given flavor has been
liked. At this point, there is no need to know how many total times
a certain flavor has been liked. It is important, however, to know
how many times a certain liked flavor has been attached to a certain
tea.